+++ Specs/Scene/Batched3DModel3DTileContentSpec.js
@@ -4,6 +4,7 @@ defineSuite([
         'Core/Color',
         'Core/HeadingPitchRange',
         'Core/HeadingPitchRoll',
+        'Core/Matrix4',
         'Core/Transforms',
         'Scene/ClippingPlane',
         'Scene/ClippingPlaneCollection',
@@ -16,6 +17,7 @@ defineSuite([
         Color,
         HeadingPitchRange,
         HeadingPitchRoll,
+        Matrix4,
         Transforms,
         ClippingPlane,
         ClippingPlaneCollection,
@@ -42,6 +44,7 @@ defineSuite([
     var deprecated1Url = './Data/Cesium3DTiles/Batched/BatchedDeprecated1/tileset.json';
     var deprecated2Url = './Data/Cesium3DTiles/Batched/BatchedDeprecated2/tileset.json';
     var gltfZUpUrl = './Data/Cesium3DTiles/Batched/BatchedGltfZUp/tileset.json';
+    var withRtcCenterUrl = './Data/Cesium3DTiles/Batched/BatchedWithRtcCenter/tileset.json';
 
     function setCamera(longitude, latitude) {
         // One feature is located at the center, point the camera there
@@ -368,6 +371,31 @@ defineSuite([
             tile.update(tileset, scene.frameState);
 
             expect(Model._getClippingFunction.calls.count()).toEqual(1);
+        });
+    });
+
+    it('transforms model positions by RTC_CENTER property in the features table', function() {
+        return Cesium3DTilesTester.loadTileset(scene, withRtcCenterUrl).then(function(tileset) {
+            Cesium3DTilesTester.expectRenderTileset(scene, tileset);
+
+            var rtcTransform = tileset._root._content._rtcCenterTransform;
+            expect(rtcTransform).toEqual(Matrix4.fromTranslation(new Cartesian3(0.1, 0.2, 0.3)));
+
+            var expectedModelTransform = Matrix4.multiply(tileset._root.transform, rtcTransform, new Matrix4());
+            expect(tileset._root._content._contentModelMatrix).toEqual(expectedModelTransform);
+            expect(tileset._root._content._model._modelMatrix).toEqual(expectedModelTransform);
+
+             // Update tile transform
+            var newLongitude = -1.31962;
+            var newLatitude = 0.698874;
+            var newCenter = Cartesian3.fromRadians(newLongitude, newLatitude, 0.0);
+            var newHPR = new HeadingPitchRoll();
+            var newTransform = Transforms.headingPitchRollToFixedFrame(newCenter, newHPR);
+            tileset._root.transform = newTransform;
+            scene.renderForSpecs();
+
+            expectedModelTransform = Matrix4.multiply(tileset._root.computedTransform, rtcTransform, expectedModelTransform);
+            expect(tileset._root._content._model._modelMatrix).toEqual(expectedModelTransform);
         });
     });